Flocking F16, Bird and Heli
experimental algorithm known as boids is an artificial life program,
developed by Craig Reynolds in 1986, using which we can simulate the
flying patterns of the bird.

The aim of the research was to simulate and understand complex system in
nature i.e. boids algorithm (flocking bird) to prove the efficiency and
open a direction to implementing it in drones.
To achieve the research aim, boids algorithm of animal movements was
simulated for different aircrafts and flock of bird, which proves the
potential of the boids algorithm and its usability in autonomous flying
drones.